Mutualize "checkSensitiveCall" function
